% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action}
\alias{Action}
\title{Construct an action object}
\usage{
Action(ID = NA, i = NA, rmVol = 0, solution = NULL, ...)
}
\arguments{
\item{ID}{a string}

\item{i}{a numeric}

\item{rmVol}{a numeric}

\item{solution}{a \code{Solution} object}

\item{...}{passed to \code{Solution}}
}
\description{
Construct an action object
}
\details{
An action object has one \code{Solution} object in the
\code{solution} slot and three other slots: \code{ID},
\code{i}, and \code{rmVol}. \code{ID} is just that, an
\code{ID} (a string) for this action. Since actions will need to be referred
to by name in many other functions, this can be very useful.
\code{i} (a numeric) is the index indicating before which data point this action
occurred. It is not the time at which the action occurred.
\code{rmVol} (a numeric) is how much volume was removed from
the well before the solution from this action was added.

If a solution object is already available, the action
can be easily created by giving it in the \code{solution} argument.
If it isn't already created, one will be created with
the \code{Solution} constructor and the \code{...} arguments.

If \code{ID}, \code{i}, or \code{rmVol} are NULL in the function call,
they are changed to the defaults in the function definition.
}

