% Generated by roxygen2 (4.0.2): do not edit by hand
\name{actionList<-}
\alias{actionList<-}
\title{Change actionsLists of well(s)}
\usage{
actionList(x) <- value
}
\arguments{
\item{x}{actionList, well, or wellList}

\item{...}{ID argument optional}
}
\description{
The only method of this S3 function is to replace the
actionList in each well in a wellList by supplying
a list of actionLists.
}

