% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actionList.actionList}
\alias{actionList.actionList}
\title{Get actionsLists from well(s)}
\usage{
actionList.actionList(x, ID = NULL, ...)
}
\arguments{
\item{x}{actionList, well, or wellList}

\item{...}{ID argument optional}
}
\description{
The default is to return the full actionList (ID=NULL)
Alternatively, an actionList containing only the last
action can be returned (ID="last")
For a wellList object, a list of actionLists are returned.
}

