% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_compounds}
\alias{add_compounds}
\title{Add compounds of two solutions}
\usage{
add_compounds(s1, s2)
}
\arguments{
\item{s1}{a \code{Solution} object}

\item{s2}{a \code{Solution} object}
}
\description{
This is the core function for adding two solutions. All compounds
are found in all of the solutions. Then the compounds are found
in each of the solutions and added together. This is iterated over
each compound.
}
\details{
The result is a data.frame of compounds, their concentrations, and types
}

