% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_data}
\alias{add_data}
\title{Add data to a well or wells}
\usage{
add_data(x, ...)
}
\arguments{
\item{wells}{wellList object}

\item{data.dir}{the root directory for the data files}

\item{parse_fun}{a function that parses the data files}
}
\description{
A wrapper that moves data from a file and parsing function
to well objects
}
\details{
\code{parse_fun} must take one input, the filepath of a
data file. It must return a data frame where the first
column is named "t" and contains the times at which data
were recorded. The other columns must have the same names
as the locations of the well. These names are used to assign
that column's data to the appropriate well in a \code{wellList} object.
}

