% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_data_data.frame}
\alias{add_data_data.frame}
\title{Add data from data frame to wells}
\usage{
add_data_data.frame(wells, df, names = FALSE)
}
\arguments{
\item{wells}{a \code{wellList} object}

\item{df}{the data frame holding the data}

\item{names}{a \code{logical}}
}
\description{
This function assumes that the column names of the data
frame are the locations of the wells. If \code{names=FALSE},
then the columns are assigned to wells in order (e.g., the third
column is assigned to the third well). The first column is assumed
to be the time. The column name "i" is reserved for the
"index" of the data points. If it is not already in the provided
data frame, then it is created as 1:nrow(df).
}

