% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_smoother}
\alias{add_smoother}
\title{Add smoother to well object}
\usage{
add_smoother(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}
}
\description{
Calls \code{smoother} and adds the result to
the well object
}

