% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_spline}
\alias{add_spline}
\title{Add a monotone Hermite spline to well}
\usage{
add_spline(x, ...)
}
\arguments{
\item{x}{a \code{well} object}

\item{...}{ignored}
}
\description{
Calls \code{stats::splinefun} with \code{method="monoH.FC"} and
adds the resulting function to the \code{well} object.
}

