% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_spline_to_plot}
\alias{add_spline_to_plot}
\title{Add spline to plot}
\usage{
add_spline_to_plot(x, nbw, min.dx = NULL, spline, discrete, maes, args, deriv)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{nbw}{the Number of points BetWeen data points to interpolate}

\item{min.dx}{related to nbw, the minimum space between points that are interpolated}

\item{spline}{\code{logical} if to interpolated data or \code{NULL}}

\item{discrete}{\code{logical} make continuous data discrete}

\item{maes}{the aes object from \code{ggplot2} to be used in the final plot}

\item{args}{the arguments from the parent \code{plot.wellList} function}

\item{deriv}{the order derivative to plot}
}
\description{
Add a spline to a plot of \code{well} objects with \code{geom_line} and the
spline functions saved within each well. Returns a \code{geom_line} layer
with it's own data.
}

