% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_title}
\alias{add_title}
\title{Add title to well plot}
\usage{
add_title(x, args, diagnostic)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{args}{the \code{args} variable in the \code{plot.wellList function}}

\item{diagnostic}{a \code{numeric} for which well to base the title on}
}
\description{
Adds a title to a well plot if there is only one
well that is being plotted. Adds the filename,
location, and concentration to the top. This is very
much an internal function.
}

