% Generated by roxygen2 (4.0.2): do not edit by hand
\name{area_under_curve}
\alias{area_under_curve}
\title{Find area under curve}
\usage{
area_under_curve(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{lower}{a \code{numeric} vector}

\item{upper}{a \code{numeric} vector}
}
\description{
Finds the area under a curve given upper and lower bounds of the x-axis.
}

