% Generated by roxygen2 (4.0.1): do not edit by hand
\name{average_replicates}
\alias{average_replicates}
\title{Average replicate wells}
\usage{
average_replicates(x, ID = "last")
}
\arguments{
\item{x}{a \code{wellList} object}

\item{ID}{which action to refer to}
}
\description{
Wells are considered replicates if their solutions
in the action specified by \code{ID} are the same.
In other words, replicates will have the same volume,
compounds and concentrations, and solvent composition.
Replicates are merged by concatenating their location
strings with a "+". Their data are merged using one
of the well's spline functions. A new \code{wellList}
object is returned. For instance, if there are 6 wells
where the are three duplicate pairs, then the output
will contain three wells.
}
\details{
The standard deviation is saved as column \code{sd} in the data
matrix of the resulting wells. \code{NA} values are removed
before the means and standard deviations are calculated.
}

