% Generated by roxygen2 (4.0.1): do not edit by hand
\name{c.actionList}
\alias{c.actionList}
\title{Concatenate actionLists}
\usage{
\method{c}{actionList}(x, ...)
}
\arguments{
\item{x}{an actionList}

\item{...}{more actionLists}
}
\description{
Custom function so that concatenation does not lose class
}

