% Generated by roxygen2 (4.0.2): do not edit by hand
\name{change_solvent}
\alias{change_solvent}
\title{Combine/subtract solvents of two solutions}
\usage{
change_solvent(s1, s2, fun)
}
\arguments{
\item{s1}{a \code{Solution} object}

\item{s2}{a \code{Solution} object}

\item{fun}{either the function \code{sum} or the function \code{subtract}}
}
\description{
Update the solvent when adding or subtracting two solutions
(e.g., \code{s1-s2} or \code{s1+s2}). The \code{fun} argument determines
if the solutions are being added or subtracted. If being added, then
set fun to \code{sum}. If being subtracted, use \code{wellz:::subtract}.
}

