% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_well_params}
\alias{check_well_params}
\title{Check well parameters}
\usage{
check_well_params(x, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{parameters and aesthetics}
}
\description{
This is principally an argument validation function
called by other functions.
It checks that the parameters that describe a well and the parameters that
will eventually be used for a plot are allowed within this framework.
}
\details{
Named arguments must be one of
\code{"color"}, \code{"linetype"}, \code{"size"}, \code{"alpha"},
\code{"shape"}, or {"fill"}. Their values must be one
\code{"location"}, \code{"file"}, \code{"volume"}, \code{"compound"},
\code{"concentration"}, \code{"solvent"}, or \code{"solvent.percentages"}.
Alternatively, the values can be an atomic vector the same length
as the number of wells in the \code{wellList}. The function aises an error
if the arguments are not valid. Otherwise, it returns \code{TRUE}.
}

