% Generated by roxygen2 (4.0.2): do not edit by hand
\name{code<-}
\alias{code<-}
\title{Set the unique identifier(s) of one or more wells}
\usage{
code(x) <- value
}
\arguments{
\item{x}{well or wellList}

\item{value}{a character vector}
}
\description{
This function sets the "code" or the string
that gives the location or identifier of a well
No two wells can have the same "code".
}

