% Generated by roxygen2 (4.0.2): do not edit by hand
\name{code}
\alias{code}
\title{Get the unique identifier(s) of one or more wells}
\usage{
code(x)
}
\arguments{
\item{x}{well or wellList}
}
\description{
This function gives the "code" or the string
that gives the location or identifier of a well
No two wells can have the same "code".
}

