% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colclasses_metadata}
\alias{colclasses_metadata}
\title{Change metadata column classes}
\usage{
colclasses_metadata(meta.df, numeric.cols = c("i", "rmVol", "adVol", "conc"))
}
\arguments{
\item{meta.df}{a \code{data.frame} of all the metadata (see \code{parse_metadata})}

\item{numeric.cols}{the names of columns to make numeric}
}
\description{
By default, the metadata is parsed in as all strings. This changes
columns with names in \code{numeric.cols} to numeric type variables.
}

