% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compound_names}
\alias{compound_names}
\title{Compound names}
\usage{
compound_names(x, ...)
}
\arguments{
\item{x}{Solution, well, or wellList}

\item{...}{ID passed to solution; unique, and collapse described below}
}
\description{
This gets the names of the compounds By default, the last
action of well is used to get the solution (ID="last"). If
ID is given but an action with that ID doesn't exist, NA is returned.
For a wellList, there are four possible types of return values
based on the unique and collapse arguments. unique=TRUE indicates
that all the compound names will be condensed into a vector of only
unique compound names. If unique=FALSE, then a list is returned (the length equal to
length(x) ). collapse=TRUE indicates that the compound names will be concatenated
together into a single string. If collapse=FALSE, a character vector is returned.
}

