% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compound_string}
\alias{compound_string}
\title{Summarize Solution compounds in one string}
\usage{
compound_string(soln, type = "all", wConc = TRUE)
}
\arguments{
\item{soln}{a \code{Solution} object}

\item{type}{the "type" of concentration, either \code{"start"},
\code{"total"}, or \code{"all"}}
}
\description{
Concatenates the names and concentrations of all compounds
}

