% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compress_data}
\alias{compress_data}
\title{Compress well data for plotting}
\usage{
compress_data(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}
}
\description{
Removes data points that couldn't possibly be displayed
because of overlap and limited screen space.
}
\details{
This function loops throught the data, removing points
that are so close to an adjacent point that they coudn't be
seen anyways. The width of the plot is found from
\code{dev.size()}. This function is fairly slow because
it uses loops in R. It wouldn't be too difficult to speed this
up with Rcpp.
}

