% Generated by roxygen2 (4.0.2): do not edit by hand
\name{concentration}
\alias{concentration}
\title{Concentration of well(s)}
\usage{
concentration(x, ...)
}
\arguments{
\item{x}{Solution, well, or wellList}

\item{...}{\code{ID} passed to solution; \code{compound} described below}
}
\description{
Returns the contents of wells depending on requested parameters (e.g., ID
or compound name).
}
\details{
For Solution objects, a compound can be requested with the
compound argument. This argument must be a character variable of length
one. If compound=NULL (the default), the concentrations
are concatenated into a string of the form
"compoundName1-12.34, compoundName2-56.7, compoundName3-8.9".
If compound does not exist in the Solution, 0 is returned.
The \code{type} argument can specify to search for compounds
of type "start" or "total" or to search "all" compounds.

For well objects, the default is to use the solution
of the last action of the well (\code{ID="last"}). \code{ID} can be specified in
to choose another solution. If \code{ID} does not specify an action,
\code{NA} is returned. \code{compound} and \code{type} can also be specified.

For wellList objects, if all the wells contain the same one compound
at different concentrations, then a numeric vector is returned.
A numeric vector is also returned if the \code{compound} is specified
as an argument. Otherwise, a vector of strings is returned as described
above for Solution objects. The last solution is used unless \code{ID}
is specified.
}

