% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagnostic_lines}
\alias{diagnostic_lines}
\title{Show data point indices on plot}
\usage{
diagnostic_lines(x, p, diagnostic)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{p}{a \code{ggplot} object to which the lines will be added}

\item{diagnostic}{a \code{numeric} indicating which well in \code{x} will
be used to make the indices and lines}
}
\description{
Add vertical lines to a plot to show the index
of a handful of points. At the thop of these lines
a the index (a number) is displayed. The lines reach
from the top of the plot down to the data curve.
}

