% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_action}
\alias{expand_action}
\title{Expand a metadata row}
\usage{
expand_action(df)
}
\arguments{
\item{df}{a \code{data.frame} holding all the rows for for one action}
}
\description{
If a row in a metadata file has shorthand for
applying the same action to several wells, this
function expands those rows are repeated for each well. The
location of the wells are expanded with \code{expand_code}.
All other columns are checked if they have comma separated values.
If there is such a column, it's values are separated among
each of the new rows.
}

