% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand_code}
\alias{expand_code}
\title{Expand well format notation}
\usage{
expand_code(code)
}
\arguments{
\item{a}{string with well location shorthand}
}
\description{
On a typical plate, wells are indicated where the rows
are letters and the columns are numbers. This code expands
shorthand. For instance, \code{"A-C1, D-E3-4, F6"} becomes
\code{c("A1","B1","C1","D3","D4","E3","E4","F6")}
}

