% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand_letters}
\alias{expand_letters}
\title{Expand ranges of letters}
\usage{
expand_letters(x)
}
\arguments{
\item{a}{string with text ranges separated by commas}
}
\description{
Expand ranges of letters. For example
\code{"A-C, D, F-H"} becomes
\code{"A","B","C","D","F","G","H"}
}

