% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_numbers}
\alias{expand_numbers}
\title{Expand ranges of numbers}
\usage{
expand_numbers(x)
}
\arguments{
\item{a}{string with number ranges separated by commas}
}
\description{
Expand ranges of numbers For example
\code{"1-4, 8, 11-13"} becomes
\code{1, 2, 3, 4, 8, 11, 12, 13)}.
}

