% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filename<-}
\alias{filename<-}
\title{Data filenames of wells}
\usage{
filename(x) <- value
}
\arguments{
\item{x}{well or wellList}

\item{value}{a character vector}
}
\description{
Sets the name of the file(s) containing
the data for each well in the input
}

