% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fillblanks_metadata}
\alias{fillblanks_metadata}
\title{fill empty cells of metadata file}
\usage{
fillblanks_metadata(x)
}
\arguments{
\item{x}{metadata file}
}
\description{
If a cell from metadata file is empty, fill it
with the value from the next non-empty cell above it.
If a cell is "-", then change it to \code{NA}.
}

