% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsmoother_composite2}
\alias{fsmoother_composite2}
\title{Variable bandwidth smoother}
\usage{
fsmoother_composite2(x, y, ...)
}
\arguments{
\item{x}{a \code{numeric} vector}

\item{y}{a \code{numeric} vector}

\item{max.knots}{a \code{numeric}}

\item{global.change}{a \code{numeric}}
}
\description{
Smoother that changes the density of knots depending
on the magnitude of change over bins of a specified width
}
\details{
The objective of this smoother is to make a smoother
that fits the data loosely when the change in the
curve is small but fit the data closely when there
are big changes and sharp spikes. The approach here
is to first bin the data and determine how many
knots will be placed in each bin. The width of the bins
are set in \code{w}. The maximum number of knots for
each bin is \code{max.knots}. The number of actual knots
in each bin equals the range of the y-values
in the bin divided by \code{global.change} times
\code{max.knots}.
}

