% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsmoother_curfit}
\alias{fsmoother_curfit}
\title{DierckxSpline curfit wrapper}
\usage{
fsmoother_curfit(x, y, w = NULL, s = NULL, knots = NULL, n = NULL, ...)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{vector of y values}

\item{w}{weights on each data point}

\item{s}{the expected sum of squared weighted residuals for
all data points (determines the amount of smoothing)}

\item{knots}{numeric vector setting location of knots}

\item{n}{the number of knots}

\item{...}{passed to \code{DierckxSpline::curfit}}
}
\description{
Makes a smoother function that returns the y-value
given and x-value. Also allows the derivative to be
returned using the \code{deriv} argument.
}

