% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsmoother_lokerns}
\alias{fsmoother_lokerns}
\title{lokerns smoother wrapper}
\usage{
fsmoother_lokerns(x, y, ...)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{vector of y values}

\item{...}{passed to \code{lokern::lokerns}}
}
\description{
Makes a smoother function that returns the y-value
given and x-value. Also allows the derivative to be
returned using the \code{deriv} argument
}

