% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsmoother_smooth.spline}
\alias{fsmoother_smooth.spline}
\title{smooth.spline wrapper}
\usage{
fsmoother_smooth.spline(x, y, ...)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{vector of y values}

\item{...}{passed to \code{stats::smooth.spline}}
}
\description{
Makes a smoother function that returns the y-value
given and x-value. Also allows the derivative to be
returned using the \code{deriv} argument
}

