% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_actionList}
\alias{get_actionList}
\title{Get actionsLists from well(s)}
\usage{
get_actionList(x, ...)
}
\arguments{
\item{x}{actionList, well, or wellList}

\item{...}{ID argument optional}
}
\description{
The default is to return the full actionList (ID=NULL)
Alternatively, an actionList containing only the last
action can be returned (ID="last")
For a wellList object, a list of actionLists are returned.
}

