% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_solution}
\alias{get_solution}
\title{Get solution(s) of a well}
\usage{
get_solution(x, ...)
}
\arguments{
\item{x}{action, actionList, well, or wellList}

\item{...}{ID argument optional, eventually passed to action}
}
\description{
For actionList and well objects, the default is to return all solutions (ID=NA)
The default for wellList objects is to return the solution of the last action
If ID is a character of length 1, then return solution for the action with that ID.
If an action with that ID doesn't exist, return NA as the solution.
Do not allow length(ID)>1.
}

