% Generated by roxygen2 (4.0.2): do not edit by hand
\name{highlight_well}
\alias{highlight_well}
\title{Highlight one curve in a plot}
\usage{
highlight_well(x, args, diagnostic)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{args}{the list of arguments in the \code{args} variable in
\code{plot.wellList}}

\item{diagnostic}{a \code{numeric} indicating which well is to be highlighted}
}
\description{
Make one curve close to opaque and all other
curves more transparent. This simply changes teh
\code{args} function in \code{plot.wellList}. It
is very much an internal function.
}

