% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ID_i}
\alias{ID_i}
\title{Get index from action name}
\usage{
ID_i(x, ...)
}
\arguments{
\item{x}{action, actionList, well, or wellList}

\item{...}{arguments passed to action or actionList}
}
\description{
Given the name of an action and a well, return the index
at which the action occurred (or right before the action occurred).
The ID of the action is passed to action or actionList with the
... arguments. If the ID isn't given,
then a named numeric vector is returned where the names are the
names of the actions and the values are the indices.
}

