% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ID_t}
\alias{ID_t}
\title{Get times from action name}
\usage{
ID_t(x, ...)
}
\arguments{
\item{x}{action, actionList, well, or wellList}

\item{...}{arguments passed to ID_i}
}
\description{
Given the name of an action and a well, return the time
at which the action occurred (or right before the action occurred).
The ID of the action is passed to action or actionList with the
... arguments. If the ID isn't given,
then a named numeric vector is returned where the names are the
names of the actions and the values are the times.
}

