% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_n_between_spline}
\alias{insert_n_between_spline}
\title{Insert points with spline or smoother}
\usage{
insert_n_between_spline(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{...}{passed to \code{insert_na_between} (see documentation for that function)}

\item{deriv}{which order derivative to return}

\item{a}{string, either \code{"spline"} or \code{"smoother"}}
}
\description{
Interpolates points in between the current data points using
the spline stored in the well object.
}

