% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_na_between}
\alias{insert_na_between}
\title{Insert NA values}
\usage{
insert_na_between(x, n, min.dx = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of NA values in between each element}

\item{the}{smallest spacing allowed between values if
the \code{NA}s were linearly interpolated}
}
\description{
Insert n NA value between each element of a numeric
vector.
}

