% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invert_list}
\alias{invert_list}
\title{Invert list names and elements}
\usage{
invert_list(x)
}
\arguments{
\item{x}{a \code{list} object}
}
\description{
Consider a list as a map.
The "keys" of a list (actually the \code{names()}) are mapped
to values. This function reverses that mapping, making a
list where the \code{names()} become the values and the values
become the \code{names()}
}

