% Generated by roxygen2 (4.0.1): do not edit by hand
\name{latex_layout}
\alias{latex_layout}
\title{Print plate layout in LaTeX}
\usage{
latex_layout(x, ID = "last", concs = TRUE, totals = FALSE, ...)
}
\arguments{
\item{x}{a \code{well} object}

\item{ID}{a \code{character}}

\item{concs}{show concentrations of wells}

\item{totals}{show compound of type "total"}

\item{...}{passed to \code{print.xtable}}
}
\description{
Print out the layout of a playout as a LaTeX table
}

