% Generated by roxygen2 (4.0.1): do not edit by hand
\name{level}
\alias{level}
\title{Level well data}
\usage{
level(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{...}{the \code{ID} can be set}
}
\description{
"Level" the data in a well to the time
of some action named \code{ID}. In other words,
all other data values will be in reference to the difference
to that point in time.
}

