% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_concat_str}
\alias{list_concat_str}
\title{summarizing a list of character vectors}
\usage{
list_concat_str(l, collapse = TRUE, unique = TRUE)
}
\arguments{
\item{l}{a \code{list} of character vectors}

\item{collapse}{a \code{logical}}

\item{unique}{a \code{logical}}
}
\description{
A very specific function, but one I use more than once
This takes a list of character vectors
\code{unique=TRUE} means that only the unique characters are returned
\code{unique=FALSE} means that the characters for each list element are returned in a vector
   the same length as the list.
\code{collapse=TRUE} means that 2+-character list elements are concatenated into one string.
\code{collapse=FASE} means that a list of character vectors is returned
}

