% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_discrete}
\alias{make_discrete}
\title{Make continous columns discrete}
\usage{
make_discrete(x, ast, discrete.params = c("colour", "fill", "size"))
}
\arguments{
\item{x}{a \code{data.table}}

\item{ast}{an \code{aes} object from \code{ggplot2}}

\item{discrete.params}{the names of columns which are allowed to be made discrete}
}
\description{
Make some columns of a \code{data.table} discrete
}
\details{
Note that this takes as input a \code{data.table}, not
\code{data.frame}. Thus, the \code{data.table} is edited
in place and does not need to be returned as is usual in R.
}

