% Generated by roxygen2 (4.0.1): do not edit by hand
\name{match_well_string}
\alias{match_well_string}
\title{Find wells from string}
\usage{
match_well_string(well, s, bounds = parse_comp_str(s), ID = "last")
}
\arguments{
\item{well}{a \code{well} object}

\item{s}{string describing desired compounds
and concentrations (see \code{search})}

\item{bounds}{an output of the same form as the output of \code{parse_comp_str(s)}}

\item{ID}{which action/Solution to use}
}
\description{
This determines if a \code{well} object matches the parameters
in a string describing desired compounds
and concentrations. See \code{search} for a description of the
required string format. The \code{ID} specifies which action and Solution
that will be used to define the well. The bounds of the concentrations
and compounds are parsed into a data frame with \code{parse_comp_str}.
See that function's documentation for more details. This function additionally
takes care of any logical operators in the string. It does this by evaluating
each compound/concentration parameter separately to \code{TRUE} or \code{FALSE}.
These logicals are then used to replace the corresponding part of the string
and the string is then evaluated as if it were regular R syntax. Because of the
regular expressions used here, one should be very careful using any
R functions. This function has only been tested with \code{|}, code{&}, and \code{!}.
}

