% Generated by roxygen2 (4.0.1): do not edit by hand
\name{max_rate}
\alias{max_rate}
\title{Find maximum rate of curve}
\usage{
max_rate(x, ...)
}
\arguments{
\item{well}{a \code{wellList} object}

\item{xlim}{a 2-element \code{numeric}}

\item{ylim}{a 2-element \code{numeric}}

\item{pgon}{a 2-column \code{data.frame}}

\item{direction}{either \code{"any"}, \code{"positive"}, or \code{"negative"}}
}
\description{
This finds the maximum derivative within boundaries
that can be defined
}
\details{
\code{xlim} and \code{ylim} define a bounding box
in which the highest derivative will be found.
\code{pgon} is a data frame where each row
represents the (x,y) coordinates of a polygon's contiguous
vertices. The derivative must be found from the data
within this polygon. If \code{direction="any"}, the largest
rate is returned regardless if it is positive or negative.
The larges positive and negative rates can be returned instead
by changin the \code{direction} argument.
}

