% Generated by roxygen2 (4.0.2): do not edit by hand
\name{melt_wellList}
\alias{melt_wellList}
\title{Melt well data}
\usage{
melt_wellList(x)
}
\arguments{
\item{x}{a \code{wellList object}}
}
\description{
Melt the data of many wells in a wellList to one
\code{data.table} where the ID variables are the
file and location of the well. The measured variables
are the times and data values for the wells.
}

