% Generated by roxygen2 (4.0.2): do not edit by hand
\name{melt_wellList_params}
\alias{melt_wellList_params}
\title{Melt a wellList with parameters}
\usage{
melt_wellList_params(x, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{passed to \code{param_matrix(x, ...)}}
}
\description{
First, the wells are melted with \code{melt_wellList}. Additional
ID variables are added to the melted \code{data.table} from
the \code{...} arguments that are passed to \code{param_matrix} (which
are then eventually passed to \code{group})
}

