% Generated by roxygen2 (4.0.1): do not edit by hand
\name{minus.Solution}
\alias{minus.Solution}
\title{Subtract a solution from another}
\usage{
minus.Solution(s1, s2, rmVol = 0)
}
\arguments{
\item{s1}{a \code{Solution} object}

\item{s2}{a \code{Solution} object}

\item{rmVol}{how much volumn is removed from \code{s1} before \code{s2} is added}
}
\description{
Realistically, a solution cannot be easily subtracted from another.
However, this function is useful if a final concentration of a solution (e.g., \code{s1}) is known
and the starting concentration is known for another solution (e.g., \code{s2})
and one wants to know what solution to add
to reach the final concentration (e.g., \code{s1-s2}).
}
\details{
Because there are different type of cencentrations, this can be tricky
(see \code{`+.Solution`}). Types of solutions that can be subtracted
(row minus column). "Y" means that the column type can be subtracted
from a solution of the type in the row.

\tabular{rcccc}{
        \tab start \tab total \tab final \tab NULL \cr
  start \tab   Y   \tab       \tab       \tab Y \cr
  total \tab       \tab   Y   \tab       \tab Y \cr
  final \tab       \tab       \tab       \tab Y \cr
  NULL  \tab       \tab       \tab       \tab
}

More error checking is done in the \code{subtract_compounds} and most
of the work is done in \code{subtract_coompound_df}.
}

