% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalize well data}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{...}{the \code{ID} can be set}
}
\description{
Normalize the data in a well to the time
of some action named \code{ID}.
}

