% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_comp_str}
\alias{parse_comp_str}
\title{Parses string for finding compounds}
\usage{
parse_comp_str(comp.string)
}
\arguments{
\item{comp.str}{see \code{search} for description of the string format}
}
\description{
See \code{search} for a description of what the
"compound string" (\code{comp.str} here) can be.
The output here is a 3-column data.frame.
First column is the name of the compound; second
is the lower bound of the concnetration; third is the
upper bound. These names and bounds are then used to search
for wells that match these requirements.
}

