% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_metadata}
\alias{parse_metadata}
\title{Parse metadata file of well actions}
\usage{
parse_metadata(metadata, data.dir = NULL, parse_fun = NULL,
  spline = !is.null(parse_fun))
}
\arguments{
\item{data.dir}{the root directory for the data files}

\item{parse_fun}{function that parses data files and returns data matrix}

\item{spline}{logical indicating to add a spline or not}
}
\description{
The metadata file must be of a csv file of a specific format. This wrapper
function calls several other functions that creates a \code{wellList} object.
}
\details{
In order, the column headers of the file must be
\code{"file"}, \code{"wells"}, \code{"ID"}, \code{"i"},
\code{"rmVol"}, \code{"adVol"}, \code{"name"}, \code{"conc"},
\code{"type"}, and \code{"solvent"}.

Empty cells of the csv file are filled with the next
non-empty cell above. The \code{wells} column is for the
location of the well. The location is a combination of a letter
(indicating the row on the plate) and a number (the column
on the plate). For example, "D3" would be the well in
fourth row and third column of a plate.
Shorthand is allowed for the \code{well} column.
For example, \code{"A-B1-2, F3-4"} would
expand to \code{c("A1","A2","B1","B2","F3","F4")} (see \code{expand_code} for
details).

The ID labels the row as an \code{action} that is being performed
right before or at the same time that data point \code{i} is collected.
\code{rmVol} and \code{adVol} are the volume removed and added, respectively,
for the action.

\code{name}, \code{conc}, and \code{type} columns refer to the compounds
in the solution added in this action. If nothing but solvent was added,
then all columns can be filled with a hyphen ("-"). Each compound must
be entered on a separate row. If there are more than two compounds in
an action (in one solution), then simply make sure that all other
columns are the same (or blank). The code will identify this as a solution
with multiple compounds. \code{type} identifies different ways that
solutions can be combined or will be tracked. \code{type="start"} means
that the given value is the concentration before the solution is added
to the well (at the start). \code{type="final"} means that this will
be the concentration of the well after \code{rmVol} is removed and
\code{adVol} is added. \code{type="total"} is the total count of some
substance in the solution (e.g., the total number of cells or beads);
a compount of type "total" is treated specially in that it's "concentration"
never changes with changing volume.

For the \code{solvent} column, only one solvent can be given
for each action. Complicated mixtures must be referred to
by a simple name (e.g., "media" or "broth").

The \code{parse_fun} argument must parse a data file and return
a data.frame where the column names are the locations of the wells, and the
first column contains the times for each data point. The name of this first
column must be "t".
}

