% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_rtca}
\alias{parse_rtca}
\title{Parse RTCA data}
\usage{
parse_rtca(filepath)
}
\arguments{
\item{filepath}{filepath of csv file exported from RTCA software}
}
\description{
A parsing function to be used by \code{add_data}. The
type of file it parses is from right clicking on the plot
in the typical RTCA xCelligence plot and saving in List format
}

