% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.well}
\alias{plot.well}
\title{Plot one well object}
\usage{
\method{plot}{well}(x, ...)
}
\arguments{
\item{x}{a \code{well} object}

\item{...}{passed to \code{plot.wellList}}
}
\description{
To plot one well object, make it a wellList and call
\code{plot.wellList}.
}

