% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.wellList}
\alias{plot.wellList}
\title{Plotting well objects}
\usage{
\method{plot}{wellList}(x, ..., diagnostic = NULL, xlim = NULL,
  points = FALSE, discrete = TRUE, replicates = FALSE, sd = replicates,
  ID = "last", spline = FALSE, line = NULL, nbw = 5, min.dx = NULL,
  smoother = FALSE, deriv = 0, compress = TRUE)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{aesthetics to add to plot}

\item{diagnostic}{a numeric indicated which well to highlight}

\item{xlim}{two-element numeric for what slice of the data to plot.
The default is to plot all data.}

\item{points}{add data points to the plot with \code{geom_point}}

\item{discrete}{aesthetics that could be continuous (e.g., concentration) are made discrete}

\item{replicates}{average wells with same solutions
with interpolation and their interpolating splines}

\item{sd}{show a ribbon plot with \code{geom_ribbon} of each replicate
group's standard deviation}

\item{ID}{the action ID passed to all other annotation functions. Wells will
be determined to be replicates based on the solutions in this action's
solution}

\item{spline}{\code{logical} for if to show the smoother function (\code{TRUE}) in the well or
just to connect the data points with straight lines (\code{FALSE}).}

\item{line}{\code{logical} for if to connect data points with lines}

\item{nbw}{the Number of points BetWeen each data point to plot so that the
smoother functions can be seen}

\item{min.dx}{related to \code{nbw}. This is the minimum spaced allowed between
points when adding interpolating points in between data points.}

\item{smoother}{\code{logical} if to show smoother}

\item{deriv}{\code{numeric} which order derivative to plot}

\item{compress}{\code{logical}, wheter to run \code{compress_data} or not}
}
\description{
A plotting function for \code{wellList} objects.
}
\details{
This sets up a \code{ggplot} object with the data from
well objects. Additional aesthetics can be set using the
\code{...} arguments. Acceptable aesthetics are
\code{"color"}, \code{"linetype"}, \code{"size"}, \code{"alpha"},
\code{"shape"}, or {"fill"}. Their values must be one
\code{"location"}, \code{"file"}, \code{"volume"}, \code{"compound"},
\code{"concentration"}, \code{"solvent"}, or \code{"solvent.percentages"}.
These aesthetics will be the names of the \code{...} arguments.
The values of these arguments will be passed to \code{group()} which
will then group the wells based on different parameters (e.g., \code{color="concentration"}
will color data from wells depending on their concentrations).

Diagnostic plots are made by setting \code{diagnostic} to a number.
For example, \code{diagnostic=2} will make the second well in the
\code{wellList} near opaque while all the other wells will be
made more transparent. Text annotation will also be added to the plot
to indicate the indices of the data points as well as the times that
are on the x-axis. The tile of the plot will give all the information
for the well of interest.
}

