% Generated by roxygen2 (4.0.1): do not edit by hand
\name{+.Solution}
\alias{+.Solution}
\title{Add two solutions}
\usage{
\method{+}{Solution}(s1, s2, rmVol = 0)
}
\arguments{
\item{s1}{a \code{Solution} object}

\item{s2}{a \code{Solution} object}

\item{rmVol}{how much volumn is removed from \code{s1} before \code{s2} is added}
}
\description{
Adds two solutions, calculating changes in concentration
and solvent composition, and returns a single \code{Solution} object.
}
\details{
The trick here is that
there are three "types" of compound concentrations. "start" means that
the solution currently has that concentration. "final" means that when
the solution is added to another solution, the concentration should be
set to this final concentration regardless of the concentration of the other
solution. "total" means that the solution has a specified number of units
and these units won't change depending on volume.

The result of adding compounds is always compound concentrations of type "start" or "total".

Types of compounds that can be added (NULL means solution doesn't have compound and
Y means that solutions can be added):

\tabular{rcccc}{
        \tab start \tab total \tab final \tab NULL \cr
  start \tab   Y   \tab       \tab   Y   \tab Y \cr
  total \tab       \tab   Y   \tab       \tab Y \cr
  final \tab   Y   \tab       \tab       \tab Y \cr
  NULL  \tab   Y   \tab   Y   \tab   Y   \tab Y
}

Most of the code to add the solutions is in \code{change_solvent} and
\code{add_compounds}.
}

