% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.Solution}
\alias{print.Solution}
\title{Print Solution object}
\usage{
\method{print}{Solution}(soln)
}
\arguments{
\item{soln}{a \code{Solution} object}
}
\description{
Print volume, compounds, solvents, and concentrations of
  a solution
}

