% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.wellList}
\alias{print.wellList}
\title{Print \code{wellList} object}
\usage{
\method{print}{wellList}(wells, printall = FALSE, ID = "last")
}
\arguments{
\item{wells}{a \code{wellList} object}

\item{printall}{print concentration of all wells even if there are >32 wells}

\item{ID}{the action and solution in each well with which to summarize the action}
}
\description{
Print the basic information of a \code{wellList} object.
If there are less than 32 wells, than a data.frame is printed
with a summary of compounds for each well
If there are more, than the number of wells in each data file
is displayed
}

