% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print_well_data}
\alias{print_well_data}
\title{Print abbreviated view of well data}
\usage{
print_well_data(data)
}
\arguments{
\item{data}{a \code{data.frame}}
}
\description{
This shows the first and last 5 rows of data so
  that the screen doesn't fill with numbers in
  other print functions. This was inspired by
  \code{print.data.table}.
}

