% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_metadata}
\alias{read_metadata}
\title{Raw parse well metadata}
\usage{
read_metadata(metadata, data.dir, sep = "\\t")
}
\arguments{
\item{metadata}{file path for the metadata file}

\item{data.dir}{directory of the data files}
}
\description{
A wrapper for \code{read.csv} for metadata files
}

