% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rollmean_fc}
\alias{rollmean_fc}
\title{A very fast rolling mean with "INTERVAL"}
\usage{
rollmean_fc(x, y, xout, width)
}
\description{
zoo::rollmean finds the "neighborhood" for the local mean by
finding the "k" closest points. This function finds all the
points that are distance "width" away from the point of interest
}

