% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sd_ribbon}
\alias{sd_ribbon}
\title{Add standard deviation ribbon}
\usage{
sd_ribbon(maes, sd)
}
\arguments{
\item{maes}{the \code{aes} object made inside \code{plot.wellList}}

\item{a}{\code{logical} if a \code{geom_ribbon} should be returned or not}
}
\description{
Make a \code{geom_ribbon} layer for \code{plot.wellList}
}

