% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select}
\alias{select}
\title{Select wells given parameters}
\usage{
select(x, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{passed to \code{search}}
}
\description{
Returns the subset of wells matching the given
parameters. The parameters are specified in the
\code{...} arguments which are passed to \code{search}.
}

