% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slice}
\alias{slice}
\title{Slice in time of well data}
\usage{
slice(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{...}{ignored}
}
\description{
Only keep data that is within the bounds
of \code{xlim}, a two-element numeric vector.
}

