% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smoother}
\alias{smoother}
\title{Make smoother function from well data}
\usage{
smoother(x, ...)
}
\arguments{
\item{x}{a \code{well} object}

\item{method}{smoother type, either \code{"smooth.spline"}
\code{"lokerns"}, \code{"curfit"}, \code{"composite"},
or \code{"composite2"}. These call
\code{fsmoother_smooth.spline}, \code{fsmoother_lokerns},
\code{fsmoother_curfit}, \code{fsmoother_composite},
and \code{fsmoother_composite2} respectively.}

\item{...}{passed to one of the fsmoother functions}
}
\description{
This returns a function that will return
a "y-value" when given an "x-value". Several
smoother functions also have an
\code{deriv} argument so that the derivative
will be returned.
}

