% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoother_maverage}
\alias{smoother_maverage}
\title{fast moving average}
\usage{
smoother_maverage(x, y, window.width)
}
\arguments{
\item{x}{the independent variable}

\item{y}{the dependent variable}

\item{window.width}{the width of the window in the moving average}
}
\description{
This function wraps an Rcpp function that calculates
the smoothing average very quickly
}

