% Generated by roxygen2 (4.0.2): do not edit by hand
\name{solution<-}
\alias{solution<-}
\title{Change solution(s) of a well}
\usage{
solution(x) <- value
}
\arguments{
\item{x}{action, actionList, or well object}

\item{value}{solution objects}
}
\description{
For actionList and well objects, the replacement value must
be a list of solutions, one for each action.
}

