% Generated by roxygen2 (4.0.2): do not edit by hand
\name{solvent_names}
\alias{solvent_names}
\title{Solvent names}
\usage{
solvent_names(x, ...)
}
\arguments{
\item{x}{Solution, well, or wellList}

\item{...}{ID passed to solution; unique, and collapse described below}
}
\description{
This gets the names of the solvents. By default, the last
action of well is used to get the solution (ID="last"). If
ID is given but an action with that ID doesn't exist, NA is returned.
For a wellList, there are four possible types of return values
based on the unique and collapse arguments. unique=TRUE indicates
that all the solvent names will be condensed into a vector of only
unique solvent names. If unique=FALSE, then a list is returned (the length equal to
length(x) ). collapse=TRUE indicates that the solvent names will be concatenated
together into a single string. If collapse=FALSE, a character vector is returned.
}

