% Generated by roxygen2 (4.0.1): do not edit by hand
\name{solvent_percentages}
\alias{solvent_percentages}
\title{Solvent composition of well(s)}
\usage{
solvent_percentages(x, ...)
}
\arguments{
\item{x}{Solution, well, or wellList}

\item{...}{\code{ID} passed to solution; \code{solvent} described below}
}
\description{
Returns the solvent percentages of wells depending on requested parameters (e.g., ID
or solvent name).
}
\details{
For Solution objects, a solvent can be requested with the
\code{solvent} argument. This argument must be a character variable of length
one. If \code{solvent=NULL} (the default), the concentrations
are concatenated into a string of the form
"solventName1-25.0, solventName2-30.0, solventName3-45.0.
If \code{solvent} does not exist in the Solution, 0 is returned.

For well objects, the default is to use the solution
of the last action of the well (\code{ID="last"}). \code{ID} can be specified in
to choose another solution. If \code{ID} does not specify an action,
\code{NA} is returned. \code{solvent} can also be specified.

For wellList objects, if all the wells contain the same one solvent
at different concentrations, then a numeric vector is returned.
A numeric vector is also returned if the \code{solvent} is specified
as an argument. Otherwise, a vector of strings is returned as described
above for \code{Solution} objects. The last solution is used unless \code{ID}
is specified.
}

