% Generated by roxygen2 (4.0.1): do not edit by hand
\name{solvent_string}
\alias{solvent_string}
\title{Summarize Solution solvents in one string}
\usage{
solvent_string(soln)
}
\arguments{
\item{soln}{a \code{Solution} object}
}
\description{
Concatenates the names and percentages of all solvents
}

