% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tcenter}
\alias{tcenter}
\title{Time-center the data}
\usage{
tcenter(x, ...)
}
\arguments{
\item{x}{a \code{well} or \code{wellList} object}

\item{...}{the \code{ID} can be set}
}
\description{
Center the data at a point identified by the action
named \code{ID}. Return a well list with the time-transformed
data.
}

