% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tdata<-}
\alias{tdata<-}
\title{Change the times for a well's data points}
\usage{
tdata(x, ...) <- value
}
\arguments{
\item{x}{well object}
}
\description{
Given a numeric vector of the times from well's data, update
the data. This is usefull for time transformations.
}

