% Generated by roxygen2 (4.0.1): do not edit by hand
\name{text_ranges}
\alias{text_ranges}
\title{String of number ranges}
\usage{
text_ranges(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Take a vector of numbers and make a string
describing them. For instance, the numbers
\code{c(1,2,3,4,10,11,15,17,18)} becomes
\code{"1-4, 10-11, 15, 17-18"}.
}

