% Generated by roxygen2 (4.0.2): do not edit by hand
\name{volume}
\alias{volume}
\title{Volume of a solution}
\usage{
volume(x, ...)
}
\arguments{
\item{x}{a well object}

\item{...}{passed to solution}
}
\description{
Returns the volume of a Solution object. If given a well
or wellList, it defaults to the last action. Otherwise, different
solutions can be specified by passing arguments to the solution
function (the ... arguments)
}

