% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wdata<-}
\alias{wdata<-}
\title{Change a well's data}
\usage{
wdata(x, ...) <- value
}
\arguments{
\item{x}{well object}
}
\description{
Change the data for a well. The matrix contains
three columns: i, t, and value indicating the indices, times, and values
}

