% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wdata}
\alias{wdata}
\title{Get a well's data}
\usage{
wdata(x, ...)
}
\arguments{
\item{x}{well object}
}
\description{
Extract the data matrix from a well. The matrix contains
three columns: i, t, and value indicating the indices, times, and values
}

