% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Well}
\alias{Well}
\title{Constructor for a well object}
\usage{
Well(file = NA, location = NA, actionList = NULL, ...)
}
\arguments{
\item{file}{}

\item{location}{}

\item{actionList}{}

\item{...}{}
}
\description{
A well object consists of the information to uniquely identify
a well, an \code{actionList} object, the data, and any data
splines or smoothers
}
\details{
The \code{file} and \code{location} arguments can be set simply
as strings.

A straightforward way to create a well object is with an already
available actionList object. If one is not available, then the approach
here is to assume that the object will have only one action (and so
one solution). Therefore, the ... arguments will be passed to
the \code{Solution} constructor. A one-action actionList will
be created, where the action will contain the solution generated
from the ... arguments and the \code{Solution} constructor.

More complicated well objects will require the more complex
Solution, action, and actionList objects to be made beforehand with their
associated constructors.
}
\examples{
# A blank well object
Well()

# A well made from an actionList
actionList = ActionList(ID=c("a","b"),
                        i=1:2,rmVol=c(0,30),
                        solutions=list(Solution(volume=90), Solution()))
Well(actionList=actionList)

# A well by passing ... to Solution()
Well(compound.names=c("A","B"),concentrations=c(10,100))
}

