% Generated by roxygen2 (4.0.1): do not edit by hand
\name{well_aes}
\alias{well_aes}
\title{aes object for wellList}
\usage{
well_aes(x, ..., type = "start", ID = "last", compound = NULL,
  solvent = NULL)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{the aesthetics}

\item{type}{ignored}

\item{ID}{ignored}

\item{compound}{ignored}

\item{solvent}{ignored}
}
\description{
Make an aes object to be used by \code{ggplot2} from the
named arguments in \code{...}.
}
\details{
The aesthetics are limited to certain types (see
\code{check_well_params}).
}

