% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WellList}
\alias{WellList}
\title{Constructor for a wellList object}
\usage{
WellList(wells = NULL, file = NULL, location = NULL, ID = NULL,
  rmVol = NULL, i = NULL, volume = NULL, compound.names = NULL,
  concentrations = NULL, compound.types = NULL, solvent.names = NULL,
  solvent.percs = NULL)
}
\description{
A \code{wellList} object is just a \code{list} of well objects.
The only purpose of having the \code{wellList} class is so that
generic functions that can be applied to \code{wellList} objects.
It also provides a nice structure so that any further
restrictions (e.g., well locations must be unique) can be
ensured.
}
\details{
Since a \code{wellList} is just a list of \code{well} objects, the easiest way
to create a \code{wellList} object is to supply a list of \code{well} objects
to the \code{wells} argument.

If well objects are not available, then this constructor only
allows for only very simple well objects to be quickly generated.
More complicated well objects with multiple actions and compounds
will have to be first made with the corresponding constructors.

If simple well objects are OK, then the way it's done is supplying
emph{n}-length vectors that will be used to create \code{Solution}
objects and one-action \code{actionList} objects for each well.
The \emph{n}-length vectors will be in the \code{file} and \code{location} (well-level parameters);
the \code{ID}, \code{rmVol}, and \code{i} (action-level parameters);
and the \code{volume}, \code{compound.names}, \code{concentrations},
\code{compound.types}, \code{solvent.names}, and \code{solvent.percs} (solution-level parameters)
arguments.
}
\examples{
# An empty wellList
WellList()

# 2-well wellList
WellList(compound.names=c("A","B"))

# 10-well wellList
WellList(location=1:10, volume=11:20)
}

