% Generated by roxygen2 (4.0.1): do not edit by hand
\name{action}
\alias{action}
\title{Get actions(s) of a well or actionList}
\usage{
action(x, ...)
}
\arguments{
\item{x}{action, actionList, well, or wellList}

\item{...}{ID argument optional}
}
\description{
The default is to return the last action (ID="last")
If ID is a character of length 1, then return action that ID.
If an action with that ID doesn't exist, return NA.
Do not allow length(ID)>1.
}

